Health checks, an ops console, and errors that group themselves
Phase 6 operations. /api/health runs two statements, executes NO DDL so polling it
never triggers a bootstrap, and returns 503 when the database is unreachable -- a
health check that answers 200 during an outage is worse than none. /api/health/deep
adds mail-transport reachability, job-queue depth, cron freshness and pool stats,
gated and failing closed because middleware hard-exempts /api/ and the route's own
check is the only thing in front of it.

/admin/ops ships ZERO client JavaScript. Refresh is a link, deliberately: a polling
dashboard is how a leaked interval exhausted the connection pool and took this whole
site down once already.

ERRORS GROUP AT WRITE TIME, not read time. trackError computes a fingerprint with
ids, emails, timestamps and bare numbers normalised, so 400 repeats of one fault
collapse to one row with a count instead of 400 rows nobody reads. Rows written
before the column existed fall back to grouping by event.

The deploy marker reads only variables Vercel already injects. Inventing a new one
would have made it wrong precisely on the emergency deploy where nobody remembers to
set it. Locally it reports known:false rather than guessing.

Indexes added where a page filters on an unindexed column -- audit_log.action was
both the filter and the facet dropdown on /admin/audit, reading a table scan every
time. Each index lives in the module that owns its table.

Runbooks under docs/ops: DEPLOYMENT, ROLLBACK, INCIDENT, BACKUP, MONITORING. Written
against this stack, not generic advice: the CRON_SECRET whitespace trap that rejects
a deploy in two seconds, the Hobby daily-cron and 100-deploy limits, and the fact
that rolling back CODE does not roll back SCHEMA because ADD COLUMN IF NOT EXISTS is
forward-only. What is NOT set up -- no external uptime monitor, no alerting channel,
no automated backup -- is stated under its own heading, because a runbook describing
monitoring nobody configured is worse than none: it is believed.

HONEST GAP: six of eight cron routes report "not instrumented" rather than a
freshness time. recordCronRun exists and is one line per route; the routes were not
edited because other workflows own several of them. Shown as its own state, never
dressed up as a failure.

astro check 193. Build completes; 47/47 pure tests pass.
NOT VERIFIED: no browser, no database. That /api/health returns 503 against a truly
unreachable database, and that pg_stat_activity is readable through the Supabase
transaction pooler, are both unproven.

+++ src/lib/analytics-domains.ts
@@ -914,7 +914,8 @@ async function payrollPanels(scope: DashboardScope, from: string, to: string): P
 
       const trend = rows(await db.execute(sql`
         SELECT make_date(year, month, 1) AS bucket,
-               CASE WHEN MAX(total_employees) >= ${MIN_GROUP} THEN COALESCE(SUM(total_net), 0)::int ELSE 0 END AS n
+               CASE WHEN MAX(total_employees) >= ${MIN_GROUP}
+                    THEN COALESCE(SUM(total_net), 0)::float8 ELSE 0 END AS n
           FROM hr_payroll_runs
          WHERE make_date(year, month, 1) >= (date_trunc('month', CURRENT_DATE) - INTERVAL '${sql.raw(String(TREND_MONTHS - 1))} months')
          GROUP BY 1 ORDER BY 1`));
@@ -1396,7 +1397,7 @@ async function financePanels(scope: DashboardScope, from: string, to: string): P
     const vendors = rows(await db.execute(sql`SELECT COUNT(*)::int AS n FROM procurement_vendors`));
 
     const trend = rows(await db.execute(sql`
-      SELECT date_trunc('month', order_date) AS bucket, COALESCE(SUM(amount), 0)::int AS n
+      SELECT date_trunc('month', order_date) AS bucket, COALESCE(SUM(amount), 0)::float8 AS n
         FROM procurement_orders
        WHERE order_date >= (date_trunc('month', CURRENT_DATE) - INTERVAL '${sql.raw(String(TREND_MONTHS - 1))} months')
        GROUP BY 1 ORDER BY 1`));
